Fix window expr deserialization
